Fixes empty directory errors.
Find spits out errors when you use "find * [options]" on an empty
directory, because it expands to "find [options]". The better choice is
"find ." as it works with the current directory only (effectively the
same). This is especially true since most of the cases that I've seen
are like:

   builtin cd "[some-path]" && find * -maxdepth 0 -type d

These should be, instead,

   builtin cd "[some-path]" && find . -maxdepth 0 -type d

This change fixes this.