Fetch more history in install-system-wide when called with --revision.
When the repo gets cloned with `git clone --depth 1` there may not be
enough history to checkout a prior commit. Since there already exists a
`--version` and `--revision` flag we can treat these 2 cases slightly
differently:

* if called with `--version` then behave as before, that is:
  `git checkout $version`
* if called with `--revision` then fetch more history before performing
  the checkout with `git fetch --depth=256`

Note that the depth of 256 is rather arbitrary but should be enough (on
average) to go back in time at least a couple of weeks.