Removed rake from global gemset. This was conflicting with rake installs in individual bundles and gemsets.
NOTE: Since a gem can have a dep on a specific version of rake, for
example, any rails install will automatically install rake (and the
specific version needed by that specific rails), and the *entire* point
of RVM is use of multiple versions of gems and rubies, its been decided
that  removing rake from at least global, preferably AND default, forces the
user to create and use fully individualised gemsets by installing rake
into that particular gemset creating the specific dependency. If they
want to use a gemset with multiple rake and rails gems installed, then a
forewarning ahead of time that that while selection can be made such
like 'rake _0.8.7_ about', it is not desirable to do it that way
BECAUSE it creates an area of unnecessary overlap where issues can and
do crop up. I can see 'default' specifying RVM's desired 'rake' version
since

aka always use a gemset, install rake into that gemset, and the only gem
that definitely stays in global AND default is the bundler gem and
anything IT needs in order to be installed should it suddenly develop
deps of its own. We only need worry here if bundler suddenly sprouts a
dep on rake of its own. bundler requires rubygems 1.8.10 at this time.
Using bundler 1.0.18 according to 'bundler --version'. Please execute
the following:  gem specification bundler -l   for more details.