Fix call to free to check memory requirements, updates #1167, fix #1166
This call was introduced by mpapis in aac1b34.
And the result use was lost while transfering requirements check from notes to
requirements in c6e79b3 (by @ddd).

In #1166, @shhhhh noticed that the check was giving an error message on OpenBSD
(and more generally, systems without `free` installed).

This commit reintroduce the lost check and fix the error message.