Cleanup Debian/Ubuntu autolibs.
After hours of @mpapis and I hating on each other we finally rewrote Debian autolibs to be better.  These changes fix: #1807 (partially -- by removing git and git-core from ruby installs, more to be done with git-core still being required by RVM,) #1799 (ping @rys -- by switching to dpkg-query which will account for virtual and meta-packages with arch,) and it allows us to group dependencies in a preferred format while still allowing people to have their own preferences by pre-installing a certain package.  The best example of this is jruby with say openjdk6 if you prefer it, even if we prefer openjdk7.

This fix also discovered #1810 and has a TODO for #1807: Finish the cleanup for ticket #1807 so we can remove git from rvm by simply detecting the type of install happening.  This will probably happen higher up maybe where we no longer run rvm requirements on a normal Ruby install?