Give clarity to code block installed to dot-files
The RVM installer edits several dot-files (and creates them, if they do
not already exist). It can be confusing for a user to find dot-files in
their home directory for shells they have never installed (e.g.
.mkshrc, if they are not a mksh user). It can also be hard for a novice
user to know which edits to their dot-files were made by what, when,
and why.

This commit fixes that confusion, by noting, for each edit:
- that the RVM installer made the edit
- when the RVM installer made the edit
- the extent of the edit, using octothorpes as delimiters. Layout
  loosely inspired by examples such as:
  https://github.com/syl20bnr/spacemacs/blob/master/core/templates/.spacemacs.template
  https://github.com/nicksp/dotfiles/blob/master/shell/bash_profile