Restore detection of gpg command for veryfication of signatures #4059
Our code can use gpg v1 without any issues to verify the signatures.
The problem is in printed instructions to download the developer keys,
for that we need to use gpg v2 as v1 is buggy and quite offten fails.

This commit restores detection of the gpg command and
fixes the instructions to use gpg2 for key download.

Update #3869, update #3544