fix support for zsh 5.4.1 onwards
Since version 5.4.1, zsh won't allow the definition of functions using
the `name()` syntax if `name` is already an alias.

The recommended way to define functions is to always use the `function`
keyword, as aliases are not expanded afterwards.

See http://zsh.sourceforge.net/releases.html (5.3.1 to 5.4.1) and
http://zsh.sourceforge.net/Doc/Release/Options.html#Scripts-and-Functions
for more details.