Capturing grep for #ruby= in variable to handle error
Capturing the return from __rmv_grep in a variable, and checking that, rather than letting it return a non-zero exit code.  The previous way would fail a bash script if set -e was enabled.   This way, we continue on failure with a non-zero exit code, but still check if we find a file with the #ruby in it.