[BUGFIX] Prevent attempted download of null versions: Honor subshell failure.
Justification for not passing subshells as an argument: Sadly, in bash it breaks
the intent of `set -o errexit`.  This example exhibits the problem:

```sh
set -o errexit

install_release()
{
  echo "Entered install_release!"
}

install_release $(exit 9)
echo 'This line is reached :('

_version=$(exit 9)
echo 'But this line is never reached :)'
install_release "$_version"
```

Outputs:
```
Entered install_release!
This line is reached :(
```