Be more specific about selecting a ruby variant
When installing or selecting a partial ruby spec like ruby-2, it should
get the latest version of that type of ruby. Instead, it was getting
the latest version of the alphabetically last type of ruby that had that
sequence of characters in its version string. The alphabetically last
ruby variant with "ruby" in its name is "truffleruby".

The problem was that the rvm_ruby_interpreter part of the version search
was missing a ^ anchor to indicate the start of the string. Adding that
anchor fixes the search.

This should also fix other occastions where you could select a partial
ruby spec, like rvm use or do.

I also fixed a similar problem with the --fuzzy option which made it
find jruby, mruby, or truffleruby, when searching for ruby. This made
rvm uninstall, reinstall and remove do the same because they use the
--fuzzy option by default.

Fixes #5114, #4986.