Avoid duplicate packet names in pkg-config
When calling pkg-config and the package is installed in two locations, `--list-all` will list the package for each instance, but we are only interested in the first instance (in preferred order of search path as defined by `PKG_CONFIG_PATH`).

Fix this by running the result through `sort -u`. This fixes a problem where otherwise broken `{C,LD,CPP,CXX}FLAGS` would be generated.