Don't warn about empty GEM_PATH
The GEM_PATH is set to an empty string by bundler when it is used per
  bundle install --path vendor/bundle

This is intentionally an checked in bundler's specs.

However rvm then shows a warning, which is very misleading, since it points to an issue, which is not present at all.
Also the warning can turn into a hard error, when the output is used from a subshell.
For instance libusb's "make install" fails then, because of "garbage" in the shell output.