Change TruffleRuby version in test
Sometimes CI tests in the macos environment fail.
The logs indicated a problem on the truffleruby side. [1]

I'm not 100% sure, but I think I was able to reproduce the problem
and eliminate it. I don't see any other option
than to upgrade the version of TruffleRuby.

The test case I changed:
> Test that the right version is installed

Instead of checking version 1.0.0-rc13, I check version 23.1.2.
The test behavior is the same because we care about the "right version".

[1]: https://github.com/oracle/truffleruby/issues/3283